YAAM: Organize feature
Organize feature added in this patch. It takes a directory and attempts
to copy all of its content, restructuring it as needed into the
destination asset directory. It tries to snap an image for each asset.
However, the images might not be exactly right depending on the model.
Some user adjustments might be needed to get the correct image.

Also the snap_image feature is added. Which simply takes a picture of an
existing asset and stores it where the asset is, to make the case above a
little easier to deal with

Signed-off-by: Amir Shehat <amir.shehata@gmail.com>